test: try to fix tests
